Add 2 new token modifiers.
The second is a simple function to make all token values lower case.
The first is more complex, and replaces the original idea for the ControllerClassName
in UTBG Issue #5 by allowing an argument to be passed to the modifier.  This argument
is the string to be used when finding text to remove from the original token text.  It
replaces the original idea for the ControllerClassName token simply by being called
with the argument "Controller" against the existing ClassName token.