Add collection of any HttpType attribute that may be on a MUT.
Adding an enum, HttpType, to the MethodDescriptor class, which holds all of the known REST
methods available to ASP.NET (even though we'll only probably use two or three of them).
Adding a function to get all of a methods Attributes.
Adding a function to look for an Http type among those attributes, and return an enum
corresponding to the first one it can find.
Will use this type in the next 2 or three tokens to generate.