Refactor WriteMethodInvocation to call new method, WriteMethodparameters
Pulled the part of the code going into the latter function that performed the appending
of the close parenthesis, and just put that right after the call.  Regardless of whether
there are any parameters, that addition has to happen.
Changed check for no parameters from an if else statement to an if that merely returns,
so that the indent of the else block's contents could be significantly decreased.