Add MethodParameters and QueryParameters tokens
Method parameters token merely calls the new WriteMethodParameters that was created
by the refactoring.  One addition was made to it, though, to only write the Method
parameters if it is Http Post or not Http at all.
Query parameters token uses a brand new function which expresses the same parameters
as they would appear in a query string.  It only does this in the opposite situation
of the Method parameters, when the Method IS an Http function that is NOT Post.