#5 Adding 3 more modifiers
AddOnParms converts parameters that would generally be the only parameters
passed to a function into parameters that are Added On to existing Parms.
Practically speaking, it merely prepends a comma to the string.
SingleLine removes all newline characters from the string, which should have
the effect of causing all parameters to a function to be placed on a single
line.
StdHttpFuncName replaces a string containing only a REST function and replaces
it with the name of a C# function that is most likely to be used to make
that kind of a REST call.  Currently only supports Get and Post.