Add new HttpMethodParameters token and adjust others
Adjusting original MethodParameters token to only give output when there
is no Http attribute on the MUT at all.
Adding the new HttpMethodParameters, which will only trigger on a Post.
When it does, it will only add the first parameter.  This does 2 things:
First, it allows the C# code to compile, since it expects only one more
parameter to be passed to the PostAsJsonAsync function after the URI string.
Second, it allows the possibility that the first parameter will be a
complex type, in which case it will need to be passed in the body of the
post operation (untested).
Adjusting the newer WriteURIParameters to only exclude writing out the
parameters if it is not an HTTP type at all, or if it is a Post and
there is only one parameter.  If there is more than one parameter, tests
have shown that, as long as the types are simple, the Post is actually
pulling its parameters from the Query string.  In that case, the parameter
passed to the C# function is ignored, so all parameters, including the
one passed to the C# function, need to be in the QueryString to be
recognized.  It is assumed, although untested, that if only one parameter
is being passed, that this will be a complex type and will be recognized
from the body, so it need not be passed in the QueryString.
Additional complexities can be imagined, but for now, this does what
the author needs it to.