#5 Refactor MethodDescriptor's new Http Property into another model class
Creating a new class, MethodAttributesDescriptor, that takes as its argument an
IEnumerable of AttributeSyntax objects, but does not traverse it until its single
Property, Http, is accessed.  When it is, it traverses the IEnumerable exactly once
until it finds an ASP.NET Http attribute, if it can find one.  It then caches and
returns the result.
The new class is instantiated and stored by the MethodDescriptor class, which
passes its constructor the IEnumerable passed to it.  By using the IEnumerable,
the loading of the AttributeSyntax objects is theoretically put off until the
information about whether the ASP.NET Http attribute exists is needed.
This could be further put off with a checkbox, unchecked by default, that the
user would have to click to turn on a flag activating Http analysis.  But that
is beyond the scope of this commit.