#5 Refactoring with rewording
Inlined a couple of out variable definitions to save space
Changed the name of SplitModifierArg to SplitModifierArgs, and changed the
arg parameter to args both in the function and the calling method to show
that the output variable could contain more than one argument.  Turns out
that the implementation doesn't look for a comma, so this is merely a
semantics change to reflect that fact.