#5 Enhance StdHttpFuncName Modifier to take arguments
Changing the function from associating http types with hard coded
function names to associating them with function names specified by
the user in the form of pairs of arguments specified in the template.
The first value is the HTTP type, and the second is the function name
to use for testing it.  Comments included to hopefully help the logic
make more sense.